requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3

# WebSocket and terminal support
websockets==12.0
ptyprocess==0.7.0

# Authentication
python-multipart==0.0.6

# Development dependencies (optional)
httpx==0.26.0  # For testing